+++ pkg/morph/client/container/put.go
@@ -1,9 +1,9 @@
 package container
 
 import (
+	"context"
 	"fmt"
 
-	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
 	fschaincontracts "github.com/nspcc-dev/neofs-node/pkg/morph/contracts"
 	"github.com/nspcc-dev/neofs-sdk-go/container"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
@@ -15,8 +15,6 @@ type PutPrm struct {
 	key   []byte
 	sig   []byte
 	token []byte
-
-	client.InvokePrmOptional
 }
 
 // SetContainer sets container.
@@ -39,27 +37,21 @@ func (p *PutPrm) SetToken(token []byte) {
 	p.token = token
 }
 
-// Put saves container with its session token, key and signature
-// in NeoFS system through Container contract call.
+// Put calls Container contract to create container with parameterized
+// credentials. If transaction is accepted for processing, Put waits for it to
+// be successfully executed. Waiting is performed within ctx,
+// [client.ErrTxAwaitTimeout] is returned when it is done.
 //
 // Returns calculated container identifier and any error
 // encountered that caused the saving to interrupt.
-func (c *Client) Put(p PutPrm) (cid.ID, error) {
+func (c *Client) Put(ctx context.Context, p PutPrm) (cid.ID, error) {
 	if len(p.sig) == 0 || len(p.key) == 0 {
 		return cid.ID{}, errNilArgument
 	}
 
-	var prm client.InvokePrm
-	prm.SetMethod(fschaincontracts.CreateContainerV2Method)
-	prm.InvokePrmOptional = p.InvokePrmOptional
-	prm.SetArgs(containerToStackItem(p.cnr), p.sig, p.key, p.token)
-
-	// no magic bugs with notary requests anymore, this operation should
-	// _always_ be notary signed so make it one more time even if it is
-	// a repeated flag setting
-	prm.RequireAlphabetSignature()
-
-	err := c.client.Invoke(prm)
+	err := c.client.CallWithAlphabetWitness(ctx, fschaincontracts.CreateContainerV2Method, []any{
+		containerToStackItem(p.cnr), p.sig, p.key, p.token,
+	})
 	if err == nil {
 		return cid.NewFromMarshalledContainer(p.cnr.Marshal()), nil
 	}
@@ -67,19 +59,20 @@ func (c *Client) Put(p PutPrm) (cid.ID, error) {
 		return cid.ID{}, fmt.Errorf("could not invoke method (%s): %w", fschaincontracts.CreateContainerV2Method, err)
 	}
 
-	prm.SetMethod(fschaincontracts.CreateContainerMethod)
-
 	domain := p.cnr.ReadDomain()
 	metaAttr := p.cnr.Attribute("__NEOFS__METAINFO_CONSISTENCY")
 	metaEnabled := metaAttr == "optimistic" || metaAttr == "strict"
 	cnrBytes := p.cnr.Marshal()
-	prm.SetArgs(cnrBytes, p.sig, p.key, p.token, domain.Name(), domain.Zone(), metaEnabled)
 
-	err = c.client.Invoke(prm)
+	err = c.client.CallWithAlphabetWitness(ctx, fschaincontracts.CreateContainerMethod, []any{
+		cnrBytes, p.sig, p.key, p.token, domain.Name(), domain.Zone(), metaEnabled,
+	})
 	if err != nil {
 		if isMethodNotFoundError(err, fschaincontracts.CreateContainerMethod) {
-			prm.SetMethod(putMethod)
-			if err = c.client.Invoke(prm); err != nil {
+			err = c.client.CallWithAlphabetWitness(ctx, putMethod, []any{
+				cnrBytes, p.sig, p.key, p.token, domain.Name(), domain.Zone(), metaEnabled,
+			})
+			if err != nil {
 				return cid.ID{}, fmt.Errorf("could not invoke method (%s): %w", putMethod, err)
 			}
 			return cid.NewFromMarshalledContainer(cnrBytes), nil